+++ app/src/main/java/com/codesoom/assignment/controllers/TaskCommandController.java
@@ -3,11 +3,8 @@ package com.codesoom.assignment.controllers;
 import com.codesoom.assignment.domain.Task;
 import com.codesoom.assignment.repository.TaskRepository;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -16,29 +13,20 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
 import java.util.Optional;
 
-@CrossOrigin
-@RequestMapping("/tasks")
+/**
+ * 할 일의 생성/수정/삭제 로직을 처리
+ */
 @RestController
-public class TaskController {
+@RequestMapping("/tasks")
+@CrossOrigin
+public class TaskCommandController {
 
     private final TaskRepository taskRepository;
 
-    public TaskController(TaskRepository taskRepository) {
+    public TaskCommandController(TaskRepository taskRepository) {
         this.taskRepository = taskRepository;
-    }
-
-    @GetMapping
-    public List<Task> getAllTasks() {
-        return taskRepository.findAll();
-    }
-
-    @GetMapping("/{id}")
-    public Task getTask(@PathVariable Long id) {
-        return taskRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException(id + "에 해당하는 Task가 존재하지 않습니다."));
     }
 
     @PostMapping
@@ -59,10 +47,5 @@ public class TaskController {
         if (deletedTask.isEmpty()) {
             throw new IllegalArgumentException(id + "에 해당하는 Task가 없어 삭제하지 못했습니다");
         }
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity handleIllegalArgumentError() {
-        return new ResponseEntity(HttpStatus.NOT_FOUND);
     }
 }
