+++ app/src/main/java/com/codesoom/assignment/controllers/TaskQueryController.java
@@ -1,0 +1,37 @@
+package com.codesoom.assignment.controllers;
+
+import com.codesoom.assignment.domain.Task;
+import com.codesoom.assignment.repository.TaskRepository;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+/**
+ * 할 일의 조회 로직을 처리
+ */
+@RestController
+@RequestMapping("/tasks")
+@CrossOrigin
+public class TaskQueryController {
+
+    private final TaskRepository taskRepository;
+
+    public TaskQueryController(TaskRepository taskRepository) {
+        this.taskRepository = taskRepository;
+    }
+
+    @GetMapping
+    public List<Task> getAllTasks() {
+        return taskRepository.findAll();
+    }
+
+    @GetMapping("/{id}")
+    public Task getTask(@PathVariable Long id) {
+        return taskRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException(id + "에 해당하는 Task가 존재하지 않습니다."));
+    }
+}
